Riley_AI/requirements.txt
pypdf==4.3.1
python-docx==1.1.2
python-pptx==0.6.23
reportlab==4.2.5

pytest==8.3.3
pytest-cov==4.1.0
ruff==0.6.8
black==24.10.0
